+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - PHPUnit version.
 - WebDriver version.
 ### Fixed
+- Check Settings value transfer.
+- Tests - WebDriver now quits even if EyesWebDriver failed to create.
 - MatchWindowTask.
 - Images tests.
 ### Added