+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Tests - WebDriver now quits even if EyesWebDriver failed to create.
 - MatchWindowTask.
 - Images tests.
+### Added
+- Eyes->getLogger() method.
 
 ## [1.3.1] - 2019-10-14
 ### Added